got rid of sets, changed integer ids to long, major refactor

+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/repository/FitnessClassRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.repository.CrudRepository;
 
 @Repository
-public interface FitnessClassRepository extends CrudRepository<FitnessClass, Integer> {
+public interface FitnessClassRepository extends CrudRepository<FitnessClass, Long> {
 }
