+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/repository/RoomRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.repository.CrudRepository;
 
 @Repository
-public interface RoomRepository extends CrudRepository<Room, Integer> {
+public interface RoomRepository extends CrudRepository<Room, Long> {
 }
