+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/repository/TraineeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.repository.CrudRepository;
 
 @Repository
-public interface TraineeRepository extends CrudRepository<Trainee, Integer> {
+public interface TraineeRepository extends CrudRepository<Trainee, Long> {
 }
