+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/service/InstructorService.java
@@ -2,5 +2,5 @@ package cz.cvut.fit.tjv.fitnessApp.service;
 
 import cz.cvut.fit.tjv.fitnessApp.domain.Instructor;
 
-public interface InstructorService extends CrudService<Instructor, Integer> {
+public interface InstructorService extends CrudService<Instructor, Long> {
 }
