+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/service/RoomService.java
@@ -2,5 +2,5 @@ package cz.cvut.fit.tjv.fitnessApp.service;
 
 import cz.cvut.fit.tjv.fitnessApp.domain.Room;
 
-public interface RoomService extends CrudService<Room, Integer> {
+public interface RoomService extends CrudService<Room, Long> {
 }
