+++ fitnessApp/src/main/java/cz/cvut/fit/tjv/fitnessApp/service/TraineeService.java
@@ -2,5 +2,5 @@ package cz.cvut.fit.tjv.fitnessApp.service;
 
 import cz.cvut.fit.tjv.fitnessApp.domain.Trainee;
 
-public interface TraineeService extends CrudService<Trainee, Integer> {
+public interface TraineeService extends CrudService<Trainee, Long> {
 }
